Memodifikasi rumus, menambahkan kolom table di beberapa database, melakuakn editing satuan, tanggal pada invoice dan masih banyak

+++ resources/views/surat_jalan/create.blade.php
@@ -566,7 +566,7 @@
                     if (satuan_jual.includes(nama_satuan)){
                         var total_jumlah = parseInt(jumlah_jual);
                     } else {
-                        var total_jumlah = parseInt(value_barang) * parseInt(jumlah_jual);
+                        var total_jumlah = parseFloat(value_barang) * parseInt(jumlah_jual);
                     }
                     var txt_total = '';
                     if(barang.includes("@")){
